requirements.txt
chardet==3.0.4
numpy==1.18.0
h5py==2.10.0
keras==2.1.5
pandas
tensorflow==1.6.0
sklearn
pillow
matplotlib
